+++ remote/src/main/java/org/ostara/remote/processor/ProcessCommand.java
@@ -3,29 +3,29 @@ package org.ostara.remote.processor;
 
 public interface ProcessCommand {
     interface Server {
-        byte SEND_MESSAGE = 1;
-        byte QUERY_TOPIC_INFOS = 2;
-        byte QUERY_CLUSTER_INFOS = 3;
-        byte REST_SUBSCRIBE = 4;
-        byte ALTER_SUBSCRIBE = 5;
-        byte CLEAN_SUBSCRIBE = 6;
-        byte CREATE_TOPIC = 7;
-        byte DELETE_TOPIC = 8;
-        byte SYNC_LEDGER = 9;
-        byte UNSYNC_LEDGER = 10;
+        int SEND_MESSAGE = 1;
+        int QUERY_TOPIC_INFOS = 2;
+        int QUERY_CLUSTER_INFOS = 3;
+        int REST_SUBSCRIBE = 4;
+        int ALTER_SUBSCRIBE = 5;
+        int CLEAN_SUBSCRIBE = 6;
+        int CREATE_TOPIC = 7;
+        int DELETE_TOPIC = 8;
+        int SYNC_LEDGER = 9;
+        int UNSYNC_LEDGER = 10;
     }
 
     interface Client {
-      byte PUSH_MESSAGE = 1;
-      byte SERVER_OFFLINE = 2;
-      byte TOPIC_INFO_CHANGED = 3;
+      int PUSH_MESSAGE = 1;
+      int SERVER_OFFLINE = 2;
+      int TOPIC_INFO_CHANGED = 3;
     }
 
     interface Failure {
-        byte UNKNOWN_EXCEPTION = -1;
-        byte TIMEOUT_EXCEPTION = -2;
-        byte CONNECT_EXCEPTION = -3;
-        byte COMMAND_EXCEPTION = -4;
-        byte PROCESS_EXCEPTION = -5;
+        int UNKNOWN_EXCEPTION = -1;
+        int TIMEOUT_EXCEPTION = -2;
+        int CONNECT_EXCEPTION = -3;
+        int COMMAND_EXCEPTION = -4;
+        int PROCESS_EXCEPTION = -5;
     }
 }
